events and wine done

+++ src/main/java/com/vmd/vmdwebshop/service/EventService.java
@@ -1,10 +1,12 @@
 package com.vmd.vmdwebshop.service;
 
 import com.vmd.vmdwebshop.Interface.AdministrativeMethodsInterface;
+import com.vmd.vmdwebshop.exception.event.*;
 import com.vmd.vmdwebshop.model.Event;
 import com.vmd.vmdwebshop.repository.EventRepository;
-import com.vmd.vmdwebshop.repository.WineRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,43 +16,86 @@ public class EventService implements AdministrativeMethodsInterface<Event> {
 
 
     private final EventRepository eventRepository;
-    private final WineRepository wineRepository;
 
     @Autowired
-    public EventService(EventRepository eventRepository, WineRepository wineRepository) {
+    public EventService(EventRepository eventRepository) {
         this.eventRepository = eventRepository;
-        this.wineRepository = wineRepository;
     }
 
+    /**
+     * This method finds and returns all existing events in the database.
+     * If no events are found, an exception will be thrown.
+     * @return list of all events in the database
+     * @exception EventNotFoundException
+     */
     @Override
     public List<Event> getAll() {
-        return eventRepository.findAll();
-    }
+        List<Event> eventList = eventRepository.findAll();
 
-
-    @Override
-    public List<Event> createAndEdit(Event event){
-
-        Event existingEvents = eventRepository.findById(event.getID());
-
-        if (existingEvents != null) {
-            existingEvents.setDate(event.getDate());
-            existingEvents.setTime(event.getTime());
-            existingEvents.setLocation(event.getLocation());
-            existingEvents.setTitle(event.getTitle());
-            existingEvents.setDescription(event.getDescription());
-            existingEvents.setImgURL(event.getImgURL());
-            existingEvents.setCancelled(event.isCancelled());
-        } else {
-            eventRepository.save(event);
+        if(eventList.isEmpty()){
+            throw new EventNotFoundException("No events were found in the database");
         }
 
+        return eventList;
+    }
+
+    /**
+     *
+     * @param event
+     * @param ID
+     * @return
+     */
+    @Override
+    public List<Event> createAndEdit(Event event, Long ID){
+
+        Event existingEvent;
+
+        try {
+            existingEvent = eventRepository.findByEventID(ID);}
+        catch (DataAccessException e) {
+            throw new EventDataAccessException("Failed to retrieve the event from the database");}
+
+        try {
+
+            if (existingEvent != null) {
+                existingEvent.setDate(event.getDate());
+                existingEvent.setTime(event.getTime());
+                existingEvent.setLocation(event.getLocation());
+                existingEvent.setTitle(event.getTitle());
+                existingEvent.setDescription(event.getDescription());
+                existingEvent.setImgURL(event.getImgURL());
+                existingEvent.setCancelled(event.isCancelled());
+                eventRepository.save(event);
+            } else {
+                eventRepository.save(event);
+            }
+
+        }catch (DataIntegrityViolationException e) {
+            throw new EventNotUpdatedException("Failed to update event");
+        }catch (DataAccessException e) {
+            throw new EventDataAccessException("Failed to save event to the database");}
+
         return eventRepository.findAll();
     }
 
+    /**
+     *
+     * @param ID
+     * @return
+     */
     @Override
     public List<Event> delete(Long ID){
-        eventRepository.deleteById(ID);
+            Event existingEvent = eventRepository.findByEventID(ID);
+
+            if(existingEvent == null){
+                throw new NullPointerException("No such event exists");
+            }
+            try {
+                eventRepository.deleteById(ID);
+            } catch (DataAccessException e) {
+                throw new EventNotDeletedException("Failed to delete event in the database");
+            }
+
         return eventRepository.findAll();
     }
 }
